removed compat lib. Moved classes so they don't clash with maven core class names
