+++ modules/maven/galasa-maven-plugin/pom.xml
@@ -90,7 +90,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.apache.maven</groupId>
-			<artifactId>maven-compat</artifactId>
+			<artifactId>maven-core</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>junit</groupId>
