+++ modules/maven/galasa-maven-plugin/src/main/java/dev/galasa/maven/plugin/galasa/BuildObrEmbeddedRepository.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.maven.plugin;
+package dev.galasa.maven.plugin.galasa;
 
 import java.io.File;
 import java.io.FileReader;
