+++ modules/maven/galasa-maven-plugin/src/test/java/dev/galasa/maven/plugin/galasa/DeployTestCatalogTest.java
@@ -3,10 +3,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.maven.plugin;
+package dev.galasa.maven.plugin.galasa;
 
 import java.net.URL;
 import org.junit.Test;
+
 import org.apache.maven.project.MavenProject;
 
 public class DeployTestCatalogTest {
