+++ modules/maven/galasa-maven-plugin/src/test/java/dev/galasa/maven/plugin/galasa/ErrorRaiserMavenImplTest.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.maven.plugin;
+package dev.galasa.maven.plugin.galasa;
 
 import static org.assertj.core.api.Assertions.*;
