+++ modules/maven/galasa-maven-plugin/src/test/java/dev/galasa/maven/plugin/galasa/MockMavenLog.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.maven.plugin;
+package dev.galasa.maven.plugin.galasa;
 import java.util.*;
 import static org.assertj.core.api.Assertions.*;
 import org.apache.maven.plugin.logging.Log;
